client/requirements.txt
ray[default]>=2.1.0, <3
requests>=2.28.0
importlib-metadata>=4.12.0
qiskit-terra>=0.21.0
qiskit-nature>=0.4.4
qiskit-ibmq-provider>=0.19.2
qiskit-aer>=0.11.1
qiskit-ibm-runtime>=0.7.0
redis>=4.3.4
